+++ hydra-node/src/Hydra/Logging/Messages.hs
@@ -10,7 +10,7 @@ module Hydra.Logging.Messages where
 
 import Hydra.Prelude
 
-import Hydra.API.Server (APIServerLog)
+import Hydra.API.APIServerLog (APIServerLog)
 import Hydra.Chain.Direct.Handlers (DirectChainLog)
 import Hydra.Network.Authenticate (AuthLog)
 import Hydra.Node (HydraNodeLog)